refactor!: move models to models package
