+++ models/rate_limits.go
@@ -1,35 +1,4 @@
-package oprish
-
-// Represents a message sent to or received from Eludris. The message contains its author and content.
-type Message struct {
-	// The author of the message, between 2-32 characters long.
-	Author string `json:"author"`
-	// The content of the message. At least one character; the upper bound is set per-instance inside Eludris.toml.
-	// This upper bound can be requested as part of the instance's InstanceInfo.
-	Content string `json:"content"`
-}
-
-// Represents info about an instance.
-type InstanceInfo struct {
-	// The name of the connected Eludris instance.
-	InstanceName string `json:"instance_name"`
-	// The description of the connected Eludris instance. If provided, this must be within 1 and 2048 characters long.
-	Description string `json:"description"`
-	// The version of the connected Eludris instance.
-	Version string `json:"version"`
-	// The maximum allowed message content length.
-	MessageLimit string `json:"message_limit"`
-	// 	The url to this instance's REST api.
-	OprishUrl string `json:"oprish_url"`
-	// The url to this instance's gateway.
-	PandemoniumUrl string `json:"pandemonium_url"`
-	// The url to this instance's CDN.
-	EffisUrl string `json:"effis_url"`
-	// The maximum number of bytes for an asset.
-	FileSize int `json:"file_size"`
-	// The maximum number of bytes for an attachment.
-	AttachmentFileSize int `json:"attachment_file_size"`
-}
+package models
 
 // Rate limits in Eludris fully depend on the instance's configuration.
 type RateLimitConf struct {
@@ -39,11 +8,11 @@ type RateLimitConf struct {
 	Limit int `json:"limit"`
 }
 
-//Represents all rate limits that apply to the connected Eludris instance.
+// Represents all rate limits that apply to the connected Eludris instance.
 // This includes individual rate limit information for Oprish (REST-api), Pandemonium (Gateway), and Effis (CDN).
 type InstanceRateLimits struct {
 	// The rate limits that apply to the connected Eludris instance's REST api.
-	Oprish OprishRateLimits  `json:"oprish"`
+	Oprish OprishRateLimits `json:"oprish"`
 	// The rate limits that apply to the connected Eludris instance's gateway.
 	Pandemonium RateLimitConf `json:"pandemonium"`
 	// The rate limits that apply to the connected Eludris instance's CDN.
